requirements.txt
streamlit
opencv-python-headless
ultralytics
numpy
pillow
